requirements.txt
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
cachetools==5.3.0
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.2
contextlib2==21.6.0
debugpy==1.6.6
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
execnet==1.9.0
executing==1.2.0
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.1.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.1.0
jupyter_core==5.3.0
knoema==1.2.46
matplotlib-inline==0.1.6
mock==5.0.1
nest-asyncio==1.5.6
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.0
pandas==1.3.2
parso==0.8.3
path==16.6.0
path.py==12.5.0
pickleshare==0.7.5
platformdirs==3.1.1
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
PyPDF2==3.0.1
pytest==7.2.2
pytest-shutil==1.7.0
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.2
termcolor==2.2.0
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
wincertstore==0.2
zipp==3.15.0